% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{ask_seafile_api_token}
\alias{ask_seafile_api_token}
\title{ask for a token}
\usage{
ask_seafile_api_token(username, password = rstudioapi::askForPassword(),
  seafile_url = get_seafile_url())
}
\arguments{
\item{username}{username}

\item{password}{password}

\item{seafile_url}{seafile base url}
}
\value{
a token as string
}
\description{
ask for a token
}
\examples{
\dontrun{
ask_seafile_api_token(seafile_url = "http://drop.legum.fr:8000/",username = "vincent@thinkr.fr")
}
}
