% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrac2.R
\name{does_path_exist}
\alias{does_path_exist}
\title{Does this path exist}
\usage{
does_path_exist(dir = "/", repos_id = get_repo_id(),
  seafile_url = get_seafile_url(), token = get_seafile_api_token())
}
\arguments{
\item{dir}{directory to techeckst}

\item{repos_id}{redos id}

\item{seafile_url}{seafile url}

\item{token}{seafile token}
}
\description{
Does this path exist
}
