% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrac2.R
\name{get_dir}
\alias{get_dir}
\title{Get directory}
\usage{
get_dir(dir = "/", repos_id = get_repo_id(),
  seafile_url = get_seafile_url(), token = get_seafile_api_token())
}
\arguments{
\item{dir}{directory}

\item{repos_id}{redos id}

\item{seafile_url}{seafile url}

\item{token}{seafile token}
}
\description{
Get directory
}
