% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_seafile_api_token}
\alias{get_seafile_api_token}
\title{get seafile api token}
\usage{
get_seafile_api_token(ask = TRUE,
  username = rstudioapi::askForPassword(prompt = "set username"),
  password = rstudioapi::askForPassword(),
  seafile_url = get_seafile_url())
}
\arguments{
\item{ask}{booleen ask if missing}

\item{username}{username}

\item{password}{password}

\item{seafile_url}{seafile base url}
}
\description{
get seafile api token
}
