% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrac2.R
\name{get_upload_link}
\alias{get_upload_link}
\title{Title}
\usage{
get_upload_link(repos_id = get_repo_id(),
  seafile_url = get_seafile_url(), token = get_seafile_api_token())
}
\arguments{
\item{repos_id}{redos id}

\item{seafile_url}{seafile url}

\item{token}{seafile token}
}
\description{
Title
}
\examples{
/dontrun{
get_upload_link(repos_id = get_repo_id(name = "Pauline"))
}
}
