% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrac.R
\name{list_libraries}
\alias{list_libraries}
\title{List libraries}
\usage{
list_libraries(seafile_url = get_seafile_url(),
  token = get_seafile_api_token())
}
\arguments{
\item{seafile_url}{seafile base url}

\item{token}{seafil token}
}
\value{
a data.frame
}
\description{
List libraries
}
