% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrac.R
\name{ping}
\alias{ping}
\title{get ping using token}
\usage{
ping(seafile_url = get_seafile_url(), token = get_seafile_api_token())
}
\arguments{
\item{seafile_url}{seafile base url}

\item{token}{token}
}
\value{
pong as string if OK
}
\description{
get ping using token
}
\examples{
ping(seafile_url = "http://drop.legum.fr:8000/",token = ask_for_token(seafile_url = "http://drop.legum.fr:8000/",username = "vincent@thinkr.fr"))
}
