% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_seafile_api_token}
\alias{set_seafile_api_token}
\title{set_seafile_api_token}
\usage{
set_seafile_api_token(token, username,
  password = rstudioapi::askForPassword(),
  seafile_url = get_seafile_url())
}
\arguments{
\item{token}{seafile api token}
}
\description{
set the seafile api token
}
