% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{ask_seafile_api_token}
\alias{ask_seafile_api_token}
\title{Ask seafile for a token}
\usage{
ask_seafile_api_token(username = rstudioapi::showPrompt(title =
  "username", message = "email"),
  password = rstudioapi::askForPassword(),
  seafile_url = get_seafile_url())
}
\arguments{
\item{username}{username}

\item{password}{password}

\item{seafile_url}{seafile base url}
}
\value{
a token as string
}
\description{
Ask seafile for a token
}
\examples{
\dontrun{
ask_seafile_api_token(seafile_url = "http://cloud.seafile.com",username = "vincent@thinkr.fr")
}
}
