% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrac2.R
\name{create_dir_r}
\alias{create_dir_r}
\title{create_dir recursive}
\usage{
create_dir_r(dir = "/", repos_id = get_repo_id(),
  seafile_url = get_seafile_url(), token = get_seafile_api_token())
}
\arguments{
\item{token}{}
}
\description{
create_dir recursive
}
