% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrac2.R
\name{get_repo_id}
\alias{get_repo_id}
\title{Title}
\usage{
get_repo_id(name = "My Library", ll = list_libraries())
}
\arguments{
\item{name}{library name}

\item{list_libraries}{all availables libraries}
}
\description{
Title
}
